导入工作台相关代码

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -62,7 +62,7 @@ public interface OrderMapper {
     @Select("select * from orders where status = #{status} and order_time < #{time}")
     List<Orders> getByStatusAndOrderTimeLT(Integer status, LocalDateTime time);
 
-    BigDecimal sumByMap(Map map);
+    Double sumByMap(Map map);
 
     Integer countByMap(Map map);
